add internal wallet creation

+++ src/modules/link/linkService.ts
@@ -149,6 +149,9 @@ export class LinkService {
 
     if (walletDoc.exists) {
       const walletData = walletDoc.data();
+      if (!!walletData.privateKey) {
+        throw new BadRequestError("You need to link another wallet before unlinking this one");
+      }
       if (walletData.userId === userId) {
         await walletDocRef.delete();
       } else {
